refactor(mappers): rename and relocate cart mappers to infra/mappers with consistent naming
Renamed and moved the following cart mapper files from `@/adapters/cart/prisma/` to `@/infra/mappers/cart/`:

- `map-domain-cart-item-to-prisma-cart-item.ts` ➜ `cart-item-entity-to-cart-item-database.mapper.ts`
- `map-domain-cart-to-prisma-cart.ts` ➜ `cart-entity-to-cart-database.mapper.ts`
- `map-domain-cart-to-prisma-db-cart.ts` ➜ `cart-entity-to-cart-database-insert.mapper.ts`
- `map-prisma-cart-item-to-domain-cart-item.ts` ➜ `cart-item-database-to-cart-item-entity.ts`
- `map-prisma-cart-to-domain-cart.ts` ➜ `cart-database-to-cart-entity.ts`

This refactor improves naming consistency across the codebase, aligns file names with the direction of data transformation, and centralizes all cart-related mappers under the `infra/mappers/cart` directory for better discoverability and maintenance.
