+++ adapters/cart/prisma/create-prisma-cart-service.adapter.ts
@@ -1,13 +1,13 @@
 import { CartEntity } from "@/domain/entities/cart.entity";
 import { createPrismaCartService } from "@/infra/services/cart/create-prisma-cart.service";
-import { mapDomainCartToPrismaCart } from "@/adapters/cart/prisma/map-domain-cart-to-prisma-cart";
-import { mapPrismaCartToDomainCart } from "@/adapters/cart/prisma/map-prisma-cart-to-domain-cart";
+import { cartEntityToCartDatabaseInsertMapper } from "@/infra/mappers/cart/cart-entity-to-cart-database-insert.mapper";
+import { mapPrismaCartToDomainCart } from "@/infra/mappers/cart/cart-database-to-cart-entity";
 
 export async function createPrismaCartServiceAdapter(
   cart: CartEntity
 ): Promise<CartEntity> {
   const response = await createPrismaCartService(
-    mapDomainCartToPrismaCart(cart)
+    cartEntityToCartDatabaseInsertMapper(cart)
   );
 
   return mapPrismaCartToDomainCart(response);
