+++ adapters/cart/prisma/update-prisma-cart-service.adapter.ts
@@ -1,9 +1,9 @@
 import { CartEntity } from "@/domain/entities/cart.entity";
 import { updatePrismaCartService } from "@/infra/services/cart/update-prisma-cart.service";
-import { mapDomainCartToPrismaDbCart } from "@/adapters/cart/prisma/map-domain-cart-to-prisma-db-cart";
+import { cartEntityToCartDatabaseMapper } from "@/infra/mappers/cart/cart-entity-to-cart-database.mapper";
 
 export async function updatePrismaCartServiceAdapter(
   cart: CartEntity
 ): Promise<void> {
-  await updatePrismaCartService(mapDomainCartToPrismaDbCart(cart));
+  await updatePrismaCartService(cartEntityToCartDatabaseMapper(cart));
 }
