+++ infra/mappers/cart/cart-database-to-cart-entity.ts
@@ -1,6 +1,6 @@
 import { CartEntity, newCartEntity } from "@/domain/entities/cart.entity";
 import { CartDatabase } from "@/infra/types/cart";
-import { mapPrismaCartItemToDomainCartItem } from "@/adapters/cart/prisma/map-prisma-cart-item-to-domain-cart-item";
+import { cartItemDatabaseToCartItemEntity } from "@/infra/mappers/cart/cart-item-database-to-cart-item-entity";
 import { newCurrencyEntity } from "@/domain/entities/currency.entity";
 import { newDateEntity } from "@/domain/entities/date.entity";
 
@@ -8,7 +8,7 @@ export function mapPrismaCartToDomainCart(dbCart: CartDatabase): CartEntity {
   return newCartEntity({
     id: dbCart.id,
     items: dbCart.items.map((dbCartItem) =>
-      mapPrismaCartItemToDomainCartItem(dbCartItem)
+      cartItemDatabaseToCartItemEntity(dbCartItem)
     ),
     createdAt: newDateEntity(dbCart.createdAt),
     itemsPrice: newCurrencyEntity(dbCart.itemsPrice),
