+++ infra/mappers/cart/cart-entity-to-cart-database-insert.mapper.ts
@@ -1,12 +1,14 @@
 import { CartEntity } from "@/domain/entities/cart.entity";
 import { CartDatabaseInsert } from "@/infra/types/cart";
-import { mapDomainCartItemToPrismaCartItem } from "@/adapters/cart/prisma/map-domain-cart-item-to-prisma-cart-item";
+import { cartItemEntityToCartItemDatabaseMapper } from "@/infra/mappers/cart/cart-item-entity-to-cart-item-database.mapper";
 
-export function mapDomainCartToPrismaCart(
+export function cartEntityToCartDatabaseInsertMapper(
   cart: CartEntity
 ): CartDatabaseInsert {
   return {
-    items: cart.items.map((item) => mapDomainCartItemToPrismaCartItem(item)),
+    items: cart.items.map((item) =>
+      cartItemEntityToCartItemDatabaseMapper(item)
+    ),
     itemsPrice: cart.itemsPrice.originalValue,
     sessionCartId: cart.sessionCartId,
     shippingPrice: cart.shippingPrice.originalValue,
