+++ infra/mappers/cart/cart-entity-to-cart-database.mapper.ts
@@ -1,11 +1,13 @@
 import { CartEntity } from "@/domain/entities/cart.entity";
 import { CartDatabase } from "@/infra/types/cart";
-import { mapDomainCartItemToPrismaCartItem } from "@/adapters/cart/prisma/map-domain-cart-item-to-prisma-cart-item";
+import { cartItemEntityToCartItemDatabaseMapper } from "@/infra/mappers/cart/cart-item-entity-to-cart-item-database.mapper";
 
-export function mapDomainCartToPrismaDbCart(cart: CartEntity): CartDatabase {
+export function cartEntityToCartDatabaseMapper(cart: CartEntity): CartDatabase {
   return {
     id: cart.id,
-    items: cart.items.map((item) => mapDomainCartItemToPrismaCartItem(item)),
+    items: cart.items.map((item) =>
+      cartItemEntityToCartItemDatabaseMapper(item)
+    ),
     itemsPrice: cart.itemsPrice.originalValue,
     sessionCartId: cart.sessionCartId,
     shippingPrice: cart.shippingPrice.originalValue,
