+++ infra/mappers/cart/cart-item-database-to-cart-item-entity.ts
@@ -2,7 +2,7 @@ import { CartItemEntity } from "@/domain/entities/cart-item.entity";
 import { newCurrencyEntity } from "@/domain/entities/currency.entity";
 import { CartItemDatabase } from "@/infra/types/cart";
 
-export function mapPrismaCartItemToDomainCartItem(
+export function cartItemDatabaseToCartItemEntity(
   dbCartItem: CartItemDatabase
 ): CartItemEntity {
   return {
